// frontend/src/api/models/PersonUpdateRequest.ts
export interface PersonUpdateRequest {
    /**
     * 
     * @type {string}
     * @memberof PersonUpdateRequest
     */
    name?: string;
    /**
     * 
     * @type {Color}
     * @memberof PersonUpdateRequest
     */
    eyeColor?: Color;
    /**
     * 
     * @type {Color}
     * @memberof PersonUpdateRequest
     */
    hairColor?: Color;
    /**
     * 
     * @type {number}
     * @memberof PersonUpdateRequest
     */
    locationId?: number;
    /**
     * 
     * @type {LocationAddRequest}
     * @memberof PersonUpdateRequest
     */
    location?: LocationAddRequest;
    /**
     * 
     * @type {boolean}
     * @memberof PersonUpdateRequest
     */
    removeLocation?: boolean;
    /**
     * 
     * @type {number}
     * @memberof PersonUpdateRequest
     */
    height?: number;
    /**
     * 
     * @type {number}
     * @memberof PersonUpdateRequest
     */
    weight?: number;
    /**
     * 
     * @type {Country}
     * @memberof PersonUpdateRequest
     */
    nationality?: Country;
}