// frontend/src/api/models/Color.ts
/* tslint:disable */
/* eslint-disable */
/**
 * IS Project API
 * Бэкенд для управления учебными группами и связанными сущностями. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const Color = {
    Black: 'BLACK',
    Yellow: 'YELLOW',
    Orange: 'ORANGE'
} as const;
export type Color = typeof Color[keyof typeof Color];


export function instanceOfColor(value: any): boolean {
    for (const key in Color) {
        if (Object.prototype.hasOwnProperty.call(Color, key)) {
            if (Color[key as keyof typeof Color] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ColorFromJSON(json: any): Color {
    return ColorFromJSONTyped(json, false);
}

export function ColorFromJSONTyped(json: any, ignoreDiscriminator: boolean): Color {
    return json as Color;
}

export function ColorToJSON(value?: Color | null): any {
    return value as any;
}

export function ColorToJSONTyped(value: any, ignoreDiscriminator: boolean): Color {
    return value as Color;
}