// frontend/src/api/models/Country.ts
/* tslint:disable */
/* eslint-disable */
/**
 * IS Project API
 * Бэкенд для управления учебными группами и связанными сущностями. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const Country = {
    UnitedKingdom: 'UNITED_KINGDOM',
    France: 'FRANCE',
    India: 'INDIA',
    Vatican: 'VATICAN'
} as const;
export type Country = typeof Country[keyof typeof Country];


export function instanceOfCountry(value: any): boolean {
    for (const key in Country) {
        if (Object.prototype.hasOwnProperty.call(Country, key)) {
            if (Country[key as keyof typeof Country] === value) {
                return true;
            }
        }
    }
    return false;
}

export function CountryFromJSON(json: any): Country {
    return CountryFromJSONTyped(json, false);
}

export function CountryFromJSONTyped(json: any, ignoreDiscriminator: boolean): Country {
    return json as Country;
}

export function CountryToJSON(value?: Country | null): any {
    return value as any;
}

export function CountryToJSONTyped(value: any, ignoreDiscriminator: boolean): Country {
    return value as Country;
}